chore: refactor into more readable modules
