+++ app-as-a-service/internal/configure-pipeline/scripts/util/manifest_helpers.go
@@ -1,0 +1,82 @@
+package util
+
+import (
+	"fmt"
+	"os"
+
+	kratix "github.com/syntasso/kratix-go"
+	appsv1 "k8s.io/api/apps/v1"
+	"sigs.k8s.io/yaml"
+)
+
+func WriteYAMLObject(sdk *kratix.KratixSDK, filename string, obj any) error {
+	b, err := yaml.Marshal(obj)
+	if err != nil {
+		return err
+	}
+	return sdk.WriteOutput(filename, b)
+}
+
+func WriteYAMLMap(sdk *kratix.KratixSDK, filename string, m map[string]any) error {
+	b, err := yaml.Marshal(m) // preserves apiVersion/kind/metadata/spec casing
+	if err != nil {
+		return err
+	}
+	return sdk.WriteOutput(filename, b)
+}
+
+func WriteServiceAccount(sdk *kratix.KratixSDK, name, namespace string) error {
+	sa := map[string]any{
+		"apiVersion": "v1",
+		"kind":       "ServiceAccount",
+		"metadata": map[string]any{
+			"name":      name,
+			"namespace": namespace,
+		},
+	}
+	return WriteYAMLMap(sdk, "serviceaccount.yaml", sa)
+}
+
+func ReadDeployment(path string) (appsv1.Deployment, error) {
+	var d appsv1.Deployment
+	b, err := os.ReadFile(path)
+	if err != nil {
+		return d, err
+	}
+	if err := yaml.Unmarshal(b, &d); err != nil {
+		return d, err
+	}
+
+	// Ensure TypeMeta present when re-serializing.
+	if d.APIVersion == "" {
+		d.APIVersion = "apps/v1"
+	}
+	if d.Kind == "" {
+		d.Kind = "Deployment"
+	}
+	return d, nil
+}
+
+func readYAMLMap(path string) (map[string]any, error) {
+	b, err := os.ReadFile(path)
+	if err != nil {
+		return nil, err
+	}
+
+	var out map[string]any
+	if err := yaml.Unmarshal(b, &out); err != nil {
+		return nil, err
+	}
+	if out == nil {
+		out = map[string]any{}
+	}
+	return out, nil
+}
+
+func WriteDestinationSelectors(ds []kratix.DestinationSelector) error {
+	data, err := yaml.Marshal(ds)
+	if err != nil {
+		return fmt.Errorf("marshal destination selectors: %w", err)
+	}
+	return os.WriteFile("/kratix/metadata/destination-selectors.yaml", data, 0o644)
+}
